% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPackages.R
\name{LoadPackages}
\alias{LoadPackages}
\title{Function for Loading a Bunch of Packages}
\usage{
LoadPackages(package_vector)
}
\arguments{
\item{package_vector}{A vector containing package names as character strings.
If no vector is provided a bunch of commonly used packages for spatial data manipulation are loaded.
Namely: "sp", "raster", "maptools", "rgdal", "rgeos", "ggmap", "snow", "randomForest", "RStoolbox",
"e1071", "kernlab", "mda", "pls", "caret", "ggplot2", "ggsn", "maps", "plyr", "reshape2", "caret".}
}
\description{
This function loads a bunch of commonly used packages for spatial data manipulation.
Alternatively the user can provide a vector containing the desired package names which should
be installed if necessary and loaded
}
\examples{
# Create vector with package names
x <- c("raster", "sp", "ggplot2")

# install if necessary and load packages
LoadPackages(package_vector = x)

# Load a bunch of commonly used packages for spatial data manipulation
LoadPackages()

}
