% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superClassAnalysis.R
\docType{package}
\name{superClassAnalysis}
\alias{superClassAnalysis}
\alias{superClassAnalysis-package}
\title{Toolbox for supervised classifications}
\description{
This R package contains several functions which are designed to increase the accurcacy
of a supervised classification. The functions are based on the superClass() function from
the RStoolbox package by Benjamin Leutner <benjamin.leutner@uni-wuerzburg.de>.
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
