% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superClassAnalysis.R
\docType{package}
\name{superClassAnalysis}
\alias{superClassAnalysis}
\alias{superClassAnalysis-package}
\title{Toolbox for supervised classifications}
\description{
This R package contains several functions which are designed to increase the accurcacy
of a supervised classification. The functions are based on the \code{\link[RStoolbox]{superClass}}
function from the \href{https://bleutner.github.io/RStoolbox/}{RStoolbox} package by Benjamin Leutner
\url{benjamin.leutner@uni-wuerzburg.de}.
}
\details{
Functions included in this package are:
\itemize{
  \item \code{\link{LoadPackages}}
  \item \code{\link{SampleSaturation}}
  \item \code{\link{ResolutionSaturation}}
  \item \code{\link{PolygonSaturation}}
}
}
\author{
Marius Philipp \email{marius.philipp@stud-mail.uni-wuerzburg.de}
}
